Move RulesForRequestingResult into a separate file
